gcp/requirements.txt
#tweepy
pandas
numpy
datetime
sqlalchemy
#pymysql
mysqlclient
#re
vaderSentiment
stop_words
python-binance
snscrape